refactor: move timestamp mixins together
This will make easier to have an overview of how timestamps are
managed in the codebase. The module shared.models.timestamps also
contains useful general comments on how to these the mixins.
