+++ src/website/shared/models/timestamps.py
@@ -12,6 +12,14 @@ different timestamp behaviours:
 from django.db import models
 
 
+class TimeStampMixin(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    class Meta:  # type: ignore[override]
+        abstract = True
+
+
 class TimeStampWithWritableUpdatedAtMixin(models.Model):
     """Timestamp mixin where `updated_at` is nullable and writable.
     Having `updated_at` being NULL at insertion allows, for example, for an easy
